feat: 表のサブグループ化をするために`csvFile`の定義を拡張した`groupedCsvFile`を作成
